removing unnecessary features
